optimize code and naming
