+++ client/kie-client/src/main/java/org/apache/servicecomb/kie/client/http/HttpRequest.java
@@ -21,27 +21,27 @@ import java.util.Map;
 
 public class HttpRequest {
 
-  private String url;
+    private String url;
 
-  private Map<String, String> headers;
+    private Map<String, String> headers;
 
-  private String content;
+    private String content;
 
-  public HttpRequest(String url, Map<String, String> headers, String content) {
-    this.url = url;
-    this.headers = headers;
-    this.content = content;
-  }
+    public HttpRequest(String url, Map<String, String> headers, String content) {
+        this.url = url;
+        this.headers = headers;
+        this.content = content;
+    }
 
-  public String getUrl() {
-    return url;
-  }
+    public String getUrl() {
+        return url;
+    }
 
-  public Map<String, String> getHeaders() {
-    return headers;
-  }
+    public Map<String, String> getHeaders() {
+        return headers;
+    }
 
-  public String getContent() {
-    return content;
-  }
+    public String getContent() {
+        return content;
+    }
 }
