+++ client/kie-client/src/main/java/org/apache/servicecomb/kie/client/http/HttpResponse.java
@@ -19,43 +19,43 @@ package org.apache.servicecomb.kie.client.http;
 
 public class HttpResponse {
 
-  private int statusCode;
+    private int statusCode;
 
-  private String message;
+    private String message;
 
-  private String content;
+    private String content;
 
-  public HttpResponse() {
+    public HttpResponse() {
 
-  }
+    }
 
-  HttpResponse(int statusCode, String message, String content) {
-    this.statusCode = statusCode;
-    this.content = content;
-    this.message = message;
-  }
+    HttpResponse(int statusCode, String message, String content) {
+        this.statusCode = statusCode;
+        this.content = content;
+        this.message = message;
+    }
 
-  public int getStatusCode() {
-    return statusCode;
-  }
+    public int getStatusCode() {
+        return statusCode;
+    }
 
-  public void setStatusCode(int statusCode) {
-    this.statusCode = statusCode;
-  }
+    public void setStatusCode(int statusCode) {
+        this.statusCode = statusCode;
+    }
 
-  public String getMessage() {
-    return message;
-  }
+    public String getMessage() {
+        return message;
+    }
 
-  public void setMessage(String message) {
-    this.message = message;
-  }
+    public void setMessage(String message) {
+        this.message = message;
+    }
 
-  public String getContent() {
-    return content;
-  }
+    public String getContent() {
+        return content;
+    }
 
-  public void setContent(String content) {
-    this.content = content;
-  }
+    public void setContent(String content) {
+        this.content = content;
+    }
 }
