+++ client/kie-client/src/main/java/org/apache/servicecomb/kie/client/http/HttpTransport.java
@@ -20,15 +20,15 @@ package org.apache.servicecomb.kie.client.http;
 import java.io.IOException;
 
 /**
- *  Created by   on 2019/10/16.
+ * Created by   on 2019/10/16.
  */
 public interface HttpTransport {
 
-  HttpResponse get(HttpRequest request) throws IOException;
+    HttpResponse get(HttpRequest request) throws IOException;
 
-  HttpResponse post(HttpRequest request) throws IOException;
+    HttpResponse post(HttpRequest request) throws IOException;
 
-  HttpResponse put(HttpRequest request) throws IOException;
+    HttpResponse put(HttpRequest request) throws IOException;
 
-  HttpResponse delete(HttpRequest request) throws IOException;
+    HttpResponse delete(HttpRequest request) throws IOException;
 }
