+++ client/kie-client/src/main/java/org/apache/servicecomb/kie/client/http/HttpTransportImpl.java
@@ -17,89 +17,100 @@
 
 package org.apache.servicecomb.kie.client.http;
 
+import org.apache.http.client.HttpClient;
 import org.apache.http.client.methods.*;
 import org.apache.http.entity.StringEntity;
-import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 import org.apache.http.util.EntityUtils;
 
 import java.io.IOException;
 
 /**
- *  Created by   on 2019/10/16.
+ * Created by   on 2019/10/16.
  */
 public class HttpTransportImpl implements HttpTransport {
 
-  private static final String HEADER_CONTENT_TYPE = "Content-Type";
+    private static final String HEADER_CONTENT_TYPE = "Content-Type";
 
-  private static final String HEADER_USER_AGENT = "User-Agent";
+    private static final String HEADER_USER_AGENT = "User-Agent";
 
-  private static final String HEADER_TENANT_NAME = "x-domain-name";
+    private static final String HEADER_TENANT_NAME = "x-domain-name";
 
-  public HttpTransportImpl() {
-  }
+    protected HttpClient httpClient;
 
-  @Override
-  public HttpResponse get(HttpRequest request) throws IOException {
-
-    HttpGet httpGet = new HttpGet(request.getUrl());
-
-    return doRequest(httpGet);
-  }
-
-  @Override
-  public HttpResponse post(HttpRequest request) throws IOException {
-
-    HttpPost httpPost = new HttpPost(request.getUrl());
-
-    if (request.getContent() != null) {
-      httpPost.setEntity(new StringEntity(request.getContent(), "UTF-8"));
-    }
-    return doRequest(httpPost);
-  }
-
-  @Override
-  public HttpResponse put(HttpRequest request) throws IOException {
-
-    HttpPut httpPut = new HttpPut(request.getUrl());
-
-    if (request.getContent() != null) {
-      httpPut.setEntity(new StringEntity(request.getContent(), "UTF-8"));
-    }
-    return doRequest(httpPut);
-  }
-
-  @Override
-  public HttpResponse delete(HttpRequest request) throws IOException {
-
-    HttpDelete httpDelete = new HttpDelete(request.getUrl());
-
-    return doRequest(httpDelete);
-  }
-
-  /**
-   * handle httpRequest
-   * @param httpRequest
-   * @return
-   * @throws IOException
-   */
-  private HttpResponse doRequest(HttpUriRequest httpRequest) throws IOException {
-    //add header
-    httpRequest.addHeader(HEADER_CONTENT_TYPE, "application/json");
-    httpRequest.addHeader(HEADER_USER_AGENT, "kie-client/1.0.0");
-    httpRequest.addHeader(HEADER_TENANT_NAME, "default");
-
-    //get Http response
-    CloseableHttpClient httpClient = HttpClients.createDefault();
-    CloseableHttpResponse response = httpClient.execute(httpRequest);
-
-    int statusCode = response.getStatusLine().getStatusCode();
-    String messgae = response.getStatusLine().getReasonPhrase();
-    String context = null;
-    if (response.getEntity() != null) {
-      context = EntityUtils.toString(response.getEntity(), "UTF-8");
+    public HttpClient getHttpClient() {
+        return httpClient;
     }
 
-    return new HttpResponse(statusCode, messgae, context);
-  }
+    public void setHttpClient(HttpClient httpClient) {
+        this.httpClient = httpClient;
+    }
+
+    public HttpTransportImpl() {
+        httpClient = HttpClients.createDefault();
+    }
+
+    @Override
+    public HttpResponse get(HttpRequest request) throws IOException {
+
+        HttpGet httpGet = new HttpGet(request.getUrl());
+
+        return doRequest(httpGet);
+    }
+
+    @Override
+    public HttpResponse post(HttpRequest request) throws IOException {
+
+        HttpPost httpPost = new HttpPost(request.getUrl());
+
+        if (request.getContent() != null) {
+            httpPost.setEntity(new StringEntity(request.getContent(), "UTF-8"));
+        }
+        return doRequest(httpPost);
+    }
+
+    @Override
+    public HttpResponse put(HttpRequest request) throws IOException {
+
+        HttpPut httpPut = new HttpPut(request.getUrl());
+
+        if (request.getContent() != null) {
+            httpPut.setEntity(new StringEntity(request.getContent(), "UTF-8"));
+        }
+        return doRequest(httpPut);
+    }
+
+    @Override
+    public HttpResponse delete(HttpRequest request) throws IOException {
+
+        HttpDelete httpDelete = new HttpDelete(request.getUrl());
+
+        return doRequest(httpDelete);
+    }
+
+    /**
+     * handle httpRequest
+     *
+     * @param httpRequest
+     * @return
+     * @throws IOException
+     */
+    private HttpResponse doRequest(HttpUriRequest httpRequest) throws IOException {
+        //add header
+        httpRequest.addHeader(HEADER_CONTENT_TYPE, "application/json");
+        httpRequest.addHeader(HEADER_USER_AGENT, "kie-client/1.0.0");
+        httpRequest.addHeader(HEADER_TENANT_NAME, "default");
+
+        //get Http response
+        org.apache.http.HttpResponse response = httpClient.execute(httpRequest);
+
+        int statusCode = response.getStatusLine().getStatusCode();
+        String messgae = response.getStatusLine().getReasonPhrase();
+        String context = null;
+        if (response.getEntity() != null) {
+            context = EntityUtils.toString(response.getEntity(), "UTF-8");
+        }
+
+        return new HttpResponse(statusCode, messgae, context);
+    }
 }
